Object cloning
To prevent original object manipulation, the value is first cloned
using a basic cloning function. The function will not handle RegExp
or any constructed object that uses private variables inside a
closure.  Possibly need a work around if an object like this must
be examined

Added a test for literal objects